PhaseRetrieval.jl Documentation

Inverse problem

Method of Gonsalves for small phases

About the package

PhaseRetrieval.jl Documentation

Manual Outline

Types and Functions

Index

Manual Outline

Forward model

Inverse problem

Types and Functions

PhaseRetrieval.camerasdict — Constant

`camerasdict` is a dictionary with often used cameras.

Use keys(camerasdict) to get the list of implemented cameras.

PhaseRetrieval.lensesdict — Constant

`lensesdict` is a dictionary with often used cameras.

Use keys(lensesdict) to get the list of implemented cameras.

PhaseRetrieval.AbstractPRproblem — Type

Big class of general phase-retrieval-related problems.

PhaseRetrieval.CameraChip — Type

CameraChip(pixelsize = p, imagesize = (sizex, sizey), bitdepth = 8, channelbitdepth = 8 repesent a camera chip with a given pixel size, imagesize, and bitdepth. If omitted, bitdepth is set to 8.

PhaseRetrieval.GS — Type

Gerchberg-Saxton, classical method to solve PR problem

PhaseRetrieval.GSparam — Type

GSparam(; ϕ⁰, maxit, maxϵ, keephistory, snapshots)

Parameters for a concrete realisation of GS method.

PhaseRetrieval.ImagingLens — Type

Fixed focal length lens

PhaseRetrieval.ImagingSensor — Type

Imaging sensor or camera consists of a lens and camera chip.

Create a sensor using keyword syntaxis (in any order), lens and cam are required

ims = ImagingSensor(lens = ::ImagingLens, cam = ::CameraChip)

Other possible keywords to simulate misalingnment

    focal_distance (= lens.focallength  by default)
    lensorigin = [0.,0.] # aka nodal point, but expressed in length units
    # misalingment parameters (not realised yet, set to zero)
    α = 0.
    β = 0.
    γ = 0.

PhaseRetrieval.Linear — Type

Build SH diversity using the linear functions

PhaseRetrieval.MLA — Type

Micro-lens array structure. 

The main reason to introduce it separable is to handle the case when camera chip is located not exactly
in the MLA focal plane.

PhaseRetrieval.PDplan — Type

PDplan(plan, diversity)

Construct Pupil Diversity plan, which, if multiplied by the array a of proper dimensions, computes $fft( diversity *a)$

PhaseRetrieval.PRproblem — Type

PRproblem(a, A)

Classical phase-retrieval problem of finding complex arrays x, X such that |x| = a, |X| = A, and X =F(x), where F denotes the Fourier transform.

PhaseRetrieval.Quadratic — Type

Build SH diversity using the quadratic functions

PhaseRetrieval.SHSensor — Type

Shack-Hartmann sensor composed from MLA and a camera. 

MLA plane position is defined by distance,  β,  and γ. 
MLA subapertures position is defined by mlaorigin and α.

PhaseRetrieval.SimConfig — Type

SimConfig(name::String, ims::PhaseRetrieval.ImagingSensor, λ::Float64) creates forward-model 
simulation environment for the PR problem obtained with  `ims` image sensor (camera + lens) at 
    wavelength `λ`. `name` is a sting identifacator used for, for instace, plotting labels.

SimConfig contains the following fields: julia name::String ims::PhaseRetrieval.ImagingSensor f::Float64 λ::Float64 d::Float64 q::Int roi::CartesianDomain2D dualroi::CartesianDomain2D ap::Array{Float64,2} mask::Array{Float64,2}`

PhaseRetrieval.hwConfig — Method

hwConfig(s::String, f, λ, d) creates a hardware configuration with
`s` camera, lens with a focal length `f` and aperture `d` and using wavelenght
`λ`.

Example

conf1 = hwConfig("UI1540", 300mm, 633nm,25mm) creates a configuration
based on UI-1540 camera, with a 1 inch lens with focal length 300mm and He-Ne laser.

AlternatingProjections.AbstractProblems.solve — Method

solve(pr::PRproblem, alg::GS, args...)

Solve PR problem using GS method. Any change of the GS parameters can be given

PhaseRetrieval.SHdiversity — Function

SHdiversity(size, cellsize, celloffset=0, α=0)

Generate SH diversity phase of given overall size, cellsize, central cell offset and rotation angle α.

PhaseRetrieval.SHdiversity — Method

SHdiversity(xrange, yrange, e₁, e₂, o)

Generate SH diversity phase on array xrange × yrange with cell based on orthogonal grid formed by vectors e₁, e₂, and origin at o.

PhaseRetrieval.appsftoPR — Method

appsftoPR(ap,psfimage) constructs a phase retrieval problem from two real arrays, representing the pupil and the focal planes intensity distributions.

The aperture and PSF are assumed to be centred in the array.

Examples

ap,_ = PhaseRetrieval.aperture(-1:.2:1,-1:.2:1,.8);
psfimage = psf(ap);

pr = appsftoPR(ap,psfimage)

# output
PRproblem{Float64, 2}([1.0 0.0 … 1.0 1.0; 0.0 0.0 … 0.0 1.0; … ; 1.0 0.0 … 1.0 1.0; 1.0 1.0 … 1.0 1.0], [8.857504209336042 4.432323038895129 … 8.857504209336042 10.878351222990858; 4.432323038895129 0.7324956483358351 … 4.432323038895129 6.182768610120748; … ; 8.857504209336042 4.432323038895129 … 8.857504209336042 10.878351222990858; 10.878351222990858 6.182768610120747 … 10.878351222990858 12.999999999999998])

PhaseRetrieval.binning — Method

binning(arr, Q)

Downsample array by replacing quadratic cells of size Q × Q by summ of its elements

PhaseRetrieval.centre — Method

Give centre of an orthogonal grid cell of the grid with cellsize, 
centre at grid origin and rotated at angle α counterclockwise.

PhaseRetrieval.constructSHPhaseDiversity — Method

constructSHPhaseDiversity(wfs::SHSensor, d, λ )

Construct SH phase diversity of minimal size for wfs within aperture of diameter d and wavelength λ. Return tuple of arrays for the aperture and phase.

PhaseRetrieval.diskmatrix — Function

Create array of zeroes with disk of ones of the relative diameter r

PhaseRetrieval.lineararray — Function

lineararray(xrange, yrange, kx, ky, k0=0)

Evaluate values of linear function $kx⋅x + ky⋅y + k0$ on array $xrange × yrange$.

lineararray(size, kx, ky, k0=0)

Use 1:size as xrange and yrange.

lineararray(xrange, yrange, a::Vector, k0=0)

Use first and second components of vector a as kx and ky.

PhaseRetrieval.logrescale — Function

logrescale(array, α = 5)

Rescale in log-scale: maximum will correspond to 1, $10^{-α}$ will correspond to 0.

PhaseRetrieval.orthIndexes — Function

Generate array of indexes of an orthogonal geometry

PhaseRetrieval.phwrap — Method

Wrap Phase

PhaseRetrieval.psf — Method

psf(amplitude, phase) -> psfimage psf(pupilfield) -> psfimage

Calculate psf for given amplitude and phase.

Examples

ap,_ = PhaseRetrieval.aperture(-1:.2:1,-1:.2:1,.8);

psfimage = psf(ap)

psfimage[1]

# output

17.720305330423212

PhaseRetrieval.quadratic — Method

Generate quadratic array

PhaseRetrieval.rescale — Method

rescale(array)

Rescale array between 0 and 1

PhaseRetrieval.rescale_minmax — Method

rescale_minmax(array)

Rescale array between 0 and 1 and return the rescaled array and the min, max values

PhaseRetrieval.slopearray — Function

slopearray(dom::CartesianDomain2D, dx, dy, k0=0) generates linear array defined on the domain such that
the total growth in x and y direction is dx and dy.

PhaseRetrieval.subdivide — Method

subdivide(arr, Q)

Subdivide array of dimension MQ x NQ in Q^2 stacked tiles of size M x N.

PhaseRetrieval.subdivide_sum — Method

subdivide_sum(arr,Q)

Divide array in quadratic cells of size Q × Q and sum the elements with the same indexes in each cell.

PhaseRetrieval.tile — Method

tile(arr, Q)

Divide array in quadratic cells of size Q × Q and stack them along the third dimension.

Index

PhaseRetrieval.camerasdict

PhaseRetrieval.lensesdict

PhaseRetrieval.AbstractPRproblem

PhaseRetrieval.CameraChip

PhaseRetrieval.GS

PhaseRetrieval.GSparam

PhaseRetrieval.ImagingLens

PhaseRetrieval.ImagingSensor

PhaseRetrieval.Linear

PhaseRetrieval.MLA

PhaseRetrieval.PDplan

PhaseRetrieval.PRproblem

PhaseRetrieval.Quadratic

PhaseRetrieval.SHSensor

PhaseRetrieval.SimConfig

PhaseRetrieval.hwConfig

AlternatingProjections.AbstractProblems.solve

PhaseRetrieval.SHdiversity

PhaseRetrieval.SHdiversity

PhaseRetrieval.appsftoPR

PhaseRetrieval.binning

PhaseRetrieval.centre

PhaseRetrieval.constructSHPhaseDiversity

PhaseRetrieval.diskmatrix

PhaseRetrieval.lineararray

PhaseRetrieval.logrescale

PhaseRetrieval.orthIndexes

PhaseRetrieval.phwrap

PhaseRetrieval.psf

PhaseRetrieval.quadratic

PhaseRetrieval.rescale

PhaseRetrieval.rescale_minmax

PhaseRetrieval.slopearray

PhaseRetrieval.subdivide

PhaseRetrieval.subdivide_sum

PhaseRetrieval.tile