PhaseRetrieval

Classic phase retrieval algorithms.

PhaseRetrieval.gerchbergsaxton

PhaseRetrieval.gerchbergsaxton — Method

gerchbergsaxton(Ain, Aout, kw...)

reconstruct the complex signals from their amplitude values based on the Gerchberg-Saxton algorithm.

References

R. W. Gerchberg and W. O. Saxton, "A practical algorithm for the determination of the phase from image and diffraction plane pictures," Optik 35, 237 (1972)