PhaseRetrieval.jl Documentation

PhaseRetrieval.jl Documentation

Types and Functions

Index

About the package

Getting started

Forward model

Initialising the simulation

Faster creation of an ImagingSensor

Quantisation and exposure level

Adding the phase aberration

Adding some diversity

Inverse problem

Using phase diversity

Hardware types

Camera chip

SimConfig

Inverse problem

Method of Gonsalves for small phases

Alternative-projections-based methods

Types and Functions

PhaseRetrieval.camerasdict — Constant

`camerasdict` is a dictionary with often used cameras.

Use keys(camerasdict) to get the list of implemented cameras.

PhaseRetrieval.lensesdict — Constant

`lensesdict` is a dictionary with often used cameras.

Use keys(lensesdict) to get the list of implemented cameras.

PhaseRetrieval.AbstractPRproblem — Type

Big class of general phase-retrieval-related problems.

PhaseRetrieval.AutoExposure — Type

AutoExposure(s =1) <: PSFExposure

Type used to set exposure of the camera to s*m, where m is the exposure that images the maximal intensity of the frame as 1. For instance, if the same scene is imaged with AutoExposure(0.5), the brightest pixel will have value 0.5. Values of s larger than 1 produce over-exposured images.

PhaseRetrieval.CameraChip — Type

CameraChip(pixelsize, imagesize; <keyword args>)

Create a camera chip with a given square pixel size and image size.

You can also specify the bit depth of the camera and the channel bit depth. For instance, 12-bit camera (bitdepth =12) often transfer its measurement packed in 16-bit number (channelbitdepth = 16).

If called on a complex field, output the image seen by the chip.

See also camerasdict, storagetype, intensity.

Arguments

pixelsize::Float64: size of the pixel. Only square pixels are supported.

imagesize::Tuple{Int, Int}: (width, height) of the image.

bitdepth::Int = 8: bit detpth of the camera.

channelbitdepth::Int = 8: channel bit depth of the camera.

Arguments can be specified in any order.

Example

cam = CameraChip(;
    pixelsize=5.3um, imagesize=(1280, 1024), bitdepth=8, channelbitdepth=8
)

julia> ap,_ = PhaseRetrieval.aperture(-1:.25:1,-1:.25:1,.8); ap
9×9 Matrix{Float64}:
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  1.0  1.0  1.0  0.0  0.0  0.0
 0.0  0.0  0.0  1.0  1.0  1.0  0.0  0.0  0.0
 0.0  0.0  0.0  1.0  1.0  1.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0

julia> imf = PhaseRetrieval.toimageplane(ap);

julia> cam = camerasdict["UI3860"];

julia> cam(imf; exposure = AutoExposure(.7)) |> collect
9×9 Array{N4f12,2} with eltype FixedPointNumbers.N4f12:
 0.0051  0.0  0.0122  0.0427  0.0601  0.0427  0.0122  0.0  0.0051
 0.0     0.0  0.0     0.0     0.0     0.0     0.0     0.0  0.0
 0.0122  0.0  0.0286  0.1006  0.1411  0.1006  0.0286  0.0  0.0122
 0.0427  0.0  0.1006  0.3553  0.4987  0.3553  0.1006  0.0  0.0427
 0.0601  0.0  0.1411  0.4987  0.6999  0.4987  0.1411  0.0  0.0601
 0.0427  0.0  0.1006  0.3553  0.4987  0.3553  0.1006  0.0  0.0427
 0.0122  0.0  0.0286  0.1006  0.1411  0.1006  0.0286  0.0  0.0122
 0.0     0.0  0.0     0.0     0.0     0.0     0.0     0.0  0.0
 0.0051  0.0  0.0122  0.0427  0.0601  0.0427  0.0122  0.0  0.0051

PhaseRetrieval.CameraChip — Method

(cam::CameraChip)(
intensity; exposure::PSFExposure=AutoExposure(), quantize=true, noise=(0, 0)

)

Calculate image generated by the camera for the given filed inensity. If used with complex array, first converts it to intesity

PhaseRetrieval.GS — Type

Gerchberg-Saxton, classical method to solve PR problem

PhaseRetrieval.GSparam — Type

GSparam(; ϕ⁰, maxit, maxϵ, keephistory, snapshots)

Parameters for a concrete realisation of GS method.

PhaseRetrieval.ImagingLens — Type

ImagingLens(;focallength, aperture)

Create fixed focal length lens with focus = focallength and aperture dimater aperture.

Arguments can be speciefied in any order.

See also diaphragm, lensesdict.

Example

julia> lens = ImagingLens(300mm, 25mm)
Imaging Lens with
  f:	300.0 mm and
  D:	25.0 mm

PhaseRetrieval.ImagingSensor — Type

ImagingSensor(lens=lens, cam=cam)

Imaging sensor or camera consisting of a lens lens and camera chip cam.

Create a sensor using keyword syntaxis (in any order), lens and cam are required

ims = ImagingSensor(lens = ::ImagingLens, cam = ::CameraChip)

See also roi, diaphragm.

Arguments:

lens::ImagingLens: lens of the camera

cam::CameraChip: sensor chip of the camera

If lens and cam are strings, corresponding values from lensesdict and camerasdict are used.

Other possible keywords to simulate misalingnment (not implemented yet)

focal_distance = lens.focallength: Distance between the lens and the sensor.

lensorigin = [0.,0.]: nodal point, expressed in length units (mm or um)

Chip plane misalingment angles:

α = 0: rotation in x

β = 0: rotation in y

γ = 0: rotation in z

Example

julia> ImagingSensor(lens = ImagingLens(300mm, 25mm),
           cam = CameraChip(
               pixelsize=5.2um,
               imagesize=(1280, 1024),
               bitdepth=8,
               channelbitdepth=8
               )
           )
Imaging Sensor made with:
  Imaging Lens with
  f:	300.0 mm and
  D:	25.0 mm and
  Camera with
	pixel size 5.2 × 5.2 μm²,
	(1280, 1024) frame and
	8/8 bit/channel output.

julia> diaphragm(ImagingSensor(lens = lensesdict["F300A25"], cam= camerasdict["UI1540"]), 10mm)
Imaging Sensor made with:
  Imaging Lens with
  f:	300.0 mm and
  D:	10.0 mm and
  Camera with
	pixel size 5.2 × 5.2 μm²,
	(1280, 1024) frame and
	8/8 bit/channel output.

julia> ImagingSensor("F300A25",  "UI1540")
Imaging Sensor made with:
  Imaging Lens with
  f:	300.0 mm and
  D:	25.0 mm and
  Camera with
	pixel size 5.2 × 5.2 μm²,
	(1280, 1024) frame and
	8/8 bit/channel output.

See also focallength, focaldistance, apdiameter.

PhaseRetrieval.LinearPolarization — Type

LinearPolarization

DOCSTRING

Fields:

orientation_angle::Float64: DESCRIPTION

PhaseRetrieval.PDPRproblem — Type

PDPRproblem(a, [Aᵢ], [ϕᵢ])

Phase-diverse phase-retrieval problem of finding complex arrays x, Xᵢ such that |x| = a, |Xᵢ| = Aᵢ, and Xᵢ =F(x exp(im ϕᵢ)), where F denotes the Fourier transform and ϕᵢ are predefined phase diversities. Note that the problem is not required to be consistent.

PDPRproblem(a, [Aᵢ], [ϕᵢ] ; center = true, renormalize = true) -> PDPRproblem

Default constructor for PDPRproblem from the arrays which origin is supposed to be at their center and array Aᵢ are renormalised to satisfy the Parseval equation.

PhaseRetrieval.PRproblem — Type

PRproblem(a, A)

Classical phase-retrieval problem of finding complex arrays x, X such that |x| = a, |X| = A, and X =F(x), where F denotes the Fourier transform. Note that the problem is not required to be consistent.

PRproblem(a,A; center = true, renormalize = true) -> PRproblem

Default constructor for PRproblem from the arrays which origin is supposed to be at their center and the second array is renormalised to satisfy the Parseval equation.

PhaseRetrieval.PRproblemSat — Type

PRproblemSat(a, A, satlevel)

Saturated PSF phase-retrieval problem of finding complex arrays x, X such that |x| = a, max(|X|, satlevel) = A, and X =F(x), where F denotes the Fourier transform.

PhaseRetrieval.ScalarComplexAmplitude — Type

ScalarComplexAmplitude <: AbstractScalarComplexAmplitude

Structure containing separately in its fields amplitude and phase.

PhaseRetrieval.ScalarPolarization — Type

ScalarPolarization

"Fake" polarization used to represent a scalar wave.

PhaseRetrieval.SimConfig — Type

SimConfig(name::String, ims::PhaseRetrieval.ImagingSensor, λ::Float64; method = PSFMethods.Fourier()) creates forward-model
simulation environment, using algorithm defined by `method`,
for the PR problem obtained with  `ims` image sensor (camera + lens) at
wavelength `λ`. `name` is a sting identifacator used for, for instace, plotting labels.

SimConfig{PSFMethods.Fourier} contains the following fields:

- name::String
- ims::PhaseRetrieval.ImagingSensor -- imaging sensor used in the configuration
- λ::Float64 -- light wavelength
- roi::CartesianDomain2D -- coordiantes in the image plane
- dualroi::CartesianDomain2D -- coordinates in the pupil plane
- ap::Array{Float64,2} -- amplitude of the pupil function
- mask::Array{Float64,2} -- array of NaNs and 1s defining the aperture support
- phases::Dict{String, PhaseBases.Phase} -- phase aberrrations present in the config
- diversity::Dict{String, PhaseBases.Phase} -- phase diversities used for PSF generation
- modulation::Dict{String, Array{Float64,2}} -- amplitude modulations used for calculation of the vectorial PSF

PhaseRetrieval.hwConfig — Method

hwConfig(s::String, f, λ, d) creates a hardware configuration with
`s` camera, lens with a focal length `f` and aperture `d` and using wavelenght
`λ`.

Example

conf1 = hwConfig("UI1540", 300mm, 633nm,25mm) creates a configuration
based on UI-1540 camera, with a 1 inch lens with focal length 300mm and He-Ne laser.

AlternatingProjections.AbstractProblems.solve — Method

solve(pr::PRproblem, alg::GS, args...)

Solve PR problem using GS method. Any change of the GS parameters can be given

PhaseRetrieval.SHdiversity — Function

SHdiversity(size, cellsize, celloffset=0, α=0)

Generate SH diversity phase of given overall size, cellsize, central cell offset and rotation angle α.

PhaseRetrieval.SHdiversity — Method

SHdiversity(xrange, yrange, e₁, e₂, o)

Generate SH diversity phase on array xrange × yrange with cell based on orthogonal grid formed by vectors e₁, e₂, and origin at o.

PhaseRetrieval.appsftoPR — Method

appsftoPR(ap,psfimage) constructs a phase retrieval problem from two real arrays, representing the pupil and the focal planes intensity distributions.

The aperture and PSF are assumed to be centred in the array.

Examples

ap,_ = PhaseRetrieval.aperture(-1:.2:1,-1:.2:1,.8);
psfimage = psf(ap);

pr = appsftoPR(ap,psfimage)

# output
PRproblem{Float64, 2}([1.0 0.0 … 1.0 1.0; 0.0 0.0 … 0.0 1.0; … ; 1.0 0.0 … 1.0 1.0; 1.0 1.0 … 1.0 1.0], [8.857504209336042 4.432323038895129 … 8.857504209336042 10.878351222990858; 4.432323038895129 0.7324956483358351 … 4.432323038895129 6.182768610120747; … ; 8.857504209336042 4.432323038895129 … 8.857504209336042 10.878351222990858; 10.878351222990858 6.182768610120746 … 10.878351222990858 12.999999999999996])

PhaseRetrieval.diaphragm — Method

diaphragm(lens, diam)

Change the lens diameter of an ImagingLens or an ImagingSensor.

Diaphragm can be larger than the lens diameter (there is no "hardware limitation").

Example

julia> diaphragm(ImagingLens(300mm, 25mm), 10mm)
Imaging Lens with
  f:	300.0 mm and
  D:	10.0 mm

PhaseRetrieval.focaldistance — Method

focaldistance(ims [,z])

Get/set the distance from the lens to the imaging plane.

PhaseRetrieval.incoherent_psf — Method

incoherent_psf(array<:Array)

Calculate 2D array corresponding to the incoherent sum of the PSF formed by slices array[:,:,ind...] for all possible indici ind.

julia> incoherent_psf(ones(3,3,2,3))
3×3 Matrix{Float64}:
 0.0    0.0  0.0
 0.0  486.0  0.0
 0.0    0.0  0.0

julia> incoherent_psf(ones(4,4,2,2,1))
4×4 Matrix{Float64}:
 0.0  0.0     0.0  0.0
 0.0  0.0     0.0  0.0
 0.0  0.0  1024.0  0.0
 0.0  0.0     0.0  0.0

julia> incoherent_psf(ones(4,4))
4×4 Matrix{Float64}:
 0.0  0.0    0.0  0.0
 0.0  0.0    0.0  0.0
 0.0  0.0  256.0  0.0
 0.0  0.0    0.0  0.0

PhaseRetrieval.logrescale — Function

logrescale(array, α = 5)

Rescale in log-scale: maximum will correspond to 1, $10^{-α}$ will correspond to 0.

PhaseRetrieval.psf — Method

psf(amplitude, phase) -> psfimage
psf(pupilfield) -> psfimage

Calculate psf for given amplitude and phase. Both PSF and pupil arrays are assumed to have the origin in their center pixel ((N+1)÷2 for dimension size N).

Examples

julia> ap,_ = PhaseRetrieval.aperture(-1:.2:1,-1:.2:1,.8); ap
11×11 Matrix{Float64}:
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  1.0  1.0  1.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  1.0  1.0  1.0  1.0  1.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  1.0  1.0  1.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0

julia> Int.(round.(psf(ap)))
11×11 Matrix{Int64}:
 18  3   1   2    0    1    0   2   1  3  18
  3  0   6   5    0    1    0   5   6  0   3
  1  6  11   4    1    5    1   4  11  6   1
  2  5   4   1   20   38   20   1   4  5   2
  0  0   1  20   78  118   78  20   1  0   0
  1  1   5  38  118  169  118  38   5  1   1
  0  0   1  20   78  118   78  20   1  0   0
  2  5   4   1   20   38   20   1   4  5   2
  1  6  11   4    1    5    1   4  11  6   1
  3  0   6   5    0    1    0   5   6  0   3
 18  3   1   2    0    1    0   2   1  3  18

julia> psf(ones(3,3),zeros(3,3))
3×3 Matrix{Float64}:
 0.0   0.0  0.0
 0.0  81.0  0.0
 0.0   0.0  0.0

julia> psf(ones(Complex{Float64},4,4))
4×4 Matrix{Float64}:
 0.0  0.0    0.0  0.0
 0.0  0.0    0.0  0.0
 0.0  0.0  256.0  0.0
 0.0  0.0    0.0  0.0

PhaseRetrieval.roi — Method

roi(cam, dims)

Create camera that represents centered regeion of interest (ROI) of cam.

ROI can be larger than the camera size! :-)

Arguments

cam: camera (CameraChip) or imaging sensor (ImagingSensor)

dims::Tuple{Int, Int}: new (width, height) of the image. Single number defines a square ROI.

PhaseRetrieval.stochasticSolve — Method

stochasticSolve(
pr::AbstractPRproblem,
algs::Tuple{Vararg{IterativeAlgorithm}},
args...;
nruns=10,
ϕscale=2π,
quality=x ->
    LinearAlgebra.norm(solution(x) .- lasty(x)) / LinearAlgebra.norm(solution(x)),
kwargs...,

)

Solve problem pr with perturbing the initial value with a random phase term uniformly distributed in interval [0,ϕscale]. Run nrun times and return solution with the minimal value of quality function.

See also stochasticSolveKeep.

PhaseRetrieval.stochasticSolveKeep — Method

stochasticSolveKeep(
pr::AbstractPRproblem,
algs::Tuple{Vararg{IterativeAlgorithm}},
args...;
nruns=10,
ϕscale=2π,
quality=x ->
    LinearAlgebra.norm(solution(x) .- lasty(x)) / LinearAlgebra.norm(solution(x)),
kwargs...,

)

Solve problem pr with perturbing the initial value with a random phase term uniformly distributed in interval [0,ϕscale]. Run nrun times and return vector of tuples of the solution and values of quality function.

See also stochasticSolve.

PhaseRetrieval.throughfocus — Method

throughfocus(conf::SimConfig, Δz)

Calculate the phase aberration (defocus) corresponding to the axial disaplcament on length Δz.

Index

PhaseRetrieval.camerasdict

PhaseRetrieval.lensesdict

PhaseRetrieval.AbstractPRproblem

PhaseRetrieval.AutoExposure

PhaseRetrieval.CameraChip

PhaseRetrieval.CameraChip

PhaseRetrieval.GS

PhaseRetrieval.GSparam

PhaseRetrieval.ImagingLens

PhaseRetrieval.ImagingSensor

PhaseRetrieval.LinearPolarization

PhaseRetrieval.PDPRproblem

PhaseRetrieval.PRproblem

PhaseRetrieval.PRproblemSat

PhaseRetrieval.ScalarComplexAmplitude

PhaseRetrieval.ScalarPolarization

PhaseRetrieval.SimConfig

PhaseRetrieval.hwConfig

AlternatingProjections.AbstractProblems.solve

PhaseRetrieval.SHdiversity

PhaseRetrieval.SHdiversity

PhaseRetrieval.appsftoPR

PhaseRetrieval.diaphragm

PhaseRetrieval.focaldistance

PhaseRetrieval.incoherent_psf

PhaseRetrieval.logrescale

PhaseRetrieval.psf

PhaseRetrieval.roi

PhaseRetrieval.stochasticSolve

PhaseRetrieval.stochasticSolveKeep

PhaseRetrieval.throughfocus