PhaseRetrieval.jl Documentation

Inverse problem

About the package

PhaseRetrieval.jl Documentation

Manual Outline

Types and Functions

Index

Inverse problem

Method of Gonsalves for small phases

Hardware types

Camera chip

Manual Outline

Inverse problem

About the package

PhaseRetrieval.jl Documentation

Manual Outline

Types and Functions

Index

Inverse problem

Method of Gonsalves for small phases

Hardware types

Camera chip

Types and Functions

PhaseRetrieval.camerasdict — Constant

`camerasdict` is a dictionary with often used cameras.

Use keys(camerasdict) to get the list of implemented cameras.

PhaseRetrieval.lensesdict — Constant

`lensesdict` is a dictionary with often used cameras.

Use keys(lensesdict) to get the list of implemented cameras.

PhaseRetrieval.storagetype — Constant

storagetype[(bitdepth,channelbitdepth)]

Get the storage type returned by the combination of the bit depth and storage depth.

PhaseRetrieval.AbstractPRproblem — Type

Big class of general phase-retrieval-related problems.

PhaseRetrieval.CameraChip — Type

CameraChip(pixelsize, imagesize; <keyword args>)

Create a camera chip with a given square pixel size and image size.

You can also specify the bit depth of the camera and the channel bit depth. For instance, 12-bit camera (bitdepth =12) often transfer its measurement packed in 16-bit number (channelbitdepth = 16).

If called on a complex field, output the image seen by the chip.

See also camerasdict, storagetype, intensity.

Arguments

pixelsize::Float64: size of the pixel. Only square pixels are supported.

imagesize::Tuple{Int, Int}: (width, height) of the image.

bitdepth::Int = 8: bit detpth of the camera.

channelbitdepth::Int = 8: channel bit depth of the camera.

Arguments can be specified in any order.

Example

cam = CameraChip(;
    pixelsize=5.3um, imagesize=(1280, 1024), bitdepth=8, channelbitdepth=8
)

julia> ap,_ = PhaseRetrieval.aperture(-1:.25:1,-1:.25:1,.8); ap
9×9 Matrix{Float64}:
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  1.0  1.0  1.0  0.0  0.0  0.0
 0.0  0.0  0.0  1.0  1.0  1.0  0.0  0.0  0.0
 0.0  0.0  0.0  1.0  1.0  1.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0

julia> imf = PhaseRetrieval.toimageplane(ap);

julia> cam = camerasdict["UI3860"];

julia> cam(imf, AutoExposure(.7)) |> collect
9×9 Array{N4f12,2} with eltype FixedPointNumbers.N4f12:
 0.0051  0.0  0.0122  0.0427  0.0601  0.0427  0.0122  0.0  0.0051
 0.0     0.0  0.0     0.0     0.0     0.0     0.0     0.0  0.0
 0.0122  0.0  0.0286  0.1006  0.1411  0.1006  0.0286  0.0  0.0122
 0.0427  0.0  0.1006  0.3553  0.4987  0.3553  0.1006  0.0  0.0427
 0.0601  0.0  0.1411  0.4987  0.6999  0.4987  0.1411  0.0  0.0601
 0.0427  0.0  0.1006  0.3553  0.4987  0.3553  0.1006  0.0  0.0427
 0.0122  0.0  0.0286  0.1006  0.1411  0.1006  0.0286  0.0  0.0122
 0.0     0.0  0.0     0.0     0.0     0.0     0.0     0.0  0.0
 0.0051  0.0  0.0122  0.0427  0.0601  0.0427  0.0122  0.0  0.0051

PhaseRetrieval.Fourier — Type

Fourier()

FFT-based method of simulation of a PSF.

PhaseRetrieval.GS — Type

Gerchberg-Saxton, classical method to solve PR problem

PhaseRetrieval.GSparam — Type

GSparam(; ϕ⁰, maxit, maxϵ, keephistory, snapshots)

Parameters for a concrete realisation of GS method.

PhaseRetrieval.ImagingLens — Type

ImagingLens(;focallength, aperture)

Create fixed focal length lens with focus = focallength and aperture dimater aperture.

Arguments can be speciefied in any order.

See also diaphragm, lensesdict.

Example

julia> lens = ImagingLens(300mm, 25mm)
ImagingLens(0.3, 0.025)

PhaseRetrieval.ImagingSensor — Type

ImagingSensor(lens=lens, cam=cam)

Imaging sensor or camera consisting of a lens lens and camera chip cam.

Create a sensor using keyword syntaxis (in any order), lens and cam are required

ims = ImagingSensor(lens = ::ImagingLens, cam = ::CameraChip)

See also roi, diaphragm.

Arguments:

lens::ImagingLens: lens of the camera

cam::CameraChip: sensor chip of the camera

If lens and cam are strings, corresponding values from lensesdict and camerasdict are used.

Other possible keywords to simulate misalingnment (not implemented yet)

focal_distance = lens.focallength: Distance between the lens and the sensor.

lensorigin = [0.,0.]: nodal point, expressed in length units (mm or um)

Chip plane misalingment angles:

α = 0: rotation in x

β = 0: rotation in y

γ = 0: rotation in z

Example

julia> ImagingSensor(lens = ImagingLens(300mm, 25mm),
           cam = CameraChip(
               pixelsize=5.2um,
               imagesize=(1280, 1024),
               bitdepth=8,
               channelbitdepth=8
               )
           )
ImagingSensor(ImagingLens(0.3, 0.025), CameraChip(5.2e-6, (1280, 1024), 8, 8), 0.3, (0.0, 0.0), 0.0, 0.0, 0.0)
 
julia> diaphragm(ImagingSensor(lens = lensesdict["F300A25"], cam= camerasdict["UI1540"]), 10mm)
ImagingSensor(ImagingLens(0.3, 0.01), CameraChip(5.2e-6, (1280, 1024), 8, 8), 0.3, (0.0, 0.0), 0.0, 0.0, 0.0)
 
julia> ImagingSensor("F300A25",  "UI1540")
ImagingSensor(ImagingLens(0.3, 0.025), CameraChip(5.2e-6, (1280, 1024), 8, 8), 0.3, (0.0, 0.0), 0.0, 0.0, 0.0)

PhaseRetrieval.Linear — Type

Build SH diversity using the linear functions

PhaseRetrieval.MLA — Type

Micro-lens array structure. 

The main reason to introduce it separable is to handle the case when camera chip is located not exactly
in the MLA focal plane.

PhaseRetrieval.PDplan — Type

PDplan(plan, diversity)

Construct Pupil Diversity plan, which, if multiplied by the array a of proper dimensions, computes $fft( diversity *a)$

PhaseRetrieval.PRproblem — Type

PRproblem(a, A)

Classical phase-retrieval problem of finding complex arrays x, X such that |x| = a, |X| = A, and X =F(x), where F denotes the Fourier transform.

PhaseRetrieval.Quadratic — Type

Build SH diversity using the quadratic functions

PhaseRetrieval.SHSensor — Type

Shack-Hartmann sensor composed from MLA and a camera. 

MLA plane position is defined by distance,  β,  and γ. 
MLA subapertures position is defined by mlaorigin and α.

PhaseRetrieval.SimConfig — Type

SimConfig(name::String, ims::PhaseRetrieval.ImagingSensor, λ::Float64) creates forward-model 
simulation environment for the PR problem obtained with  `ims` image sensor (camera + lens) at 
    wavelength `λ`. `name` is a sting identifacator used for, for instace, plotting labels.

SimConfig contains the following fields: julia name::String ims::PhaseRetrieval.ImagingSensor f::Float64 λ::Float64 d::Float64 q::Int roi::CartesianDomain2D dualroi::CartesianDomain2D ap::Array{Float64,2} mask::Array{Float64,2}`

PhaseRetrieval.hwConfig — Method

hwConfig(s::String, f, λ, d) creates a hardware configuration with
`s` camera, lens with a focal length `f` and aperture `d` and using wavelenght
`λ`.

Example

conf1 = hwConfig("UI1540", 300mm, 633nm,25mm) creates a configuration
based on UI-1540 camera, with a 1 inch lens with focal length 300mm and He-Ne laser.

AlternatingProjections.AbstractProblems.solve — Method

solve(pr::PRproblem, alg::GS, args...)

Solve PR problem using GS method. Any change of the GS parameters can be given

PhaseRetrieval.SHdiversity — Function

SHdiversity(size, cellsize, celloffset=0, α=0)

Generate SH diversity phase of given overall size, cellsize, central cell offset and rotation angle α.

PhaseRetrieval.SHdiversity — Method

SHdiversity(xrange, yrange, e₁, e₂, o)

Generate SH diversity phase on array xrange × yrange with cell based on orthogonal grid formed by vectors e₁, e₂, and origin at o.

PhaseRetrieval.appsftoPR — Method

appsftoPR(ap,psfimage) constructs a phase retrieval problem from two real arrays, representing the pupil and the focal planes intensity distributions.

The aperture and PSF are assumed to be centred in the array.

Examples

ap,_ = PhaseRetrieval.aperture(-1:.2:1,-1:.2:1,.8);
psfimage = psf(ap);

pr = appsftoPR(ap,psfimage)

# output
PRproblem{Float64, 2}([1.0 0.0 … 1.0 1.0; 0.0 0.0 … 0.0 1.0; … ; 1.0 0.0 … 1.0 1.0; 1.0 1.0 … 1.0 1.0], [8.857504209336042 4.432323038895129 … 8.857504209336042 10.878351222990858; 4.432323038895129 0.7324956483358351 … 4.432323038895129 6.182768610120748; … ; 8.857504209336042 4.432323038895129 … 8.857504209336042 10.878351222990858; 10.878351222990858 6.182768610120747 … 10.878351222990858 12.999999999999998])

PhaseRetrieval.binning — Method

binning(arr, Q)

Downsample array by replacing quadratic cells of size Q × Q by summ of its elements

PhaseRetrieval.centre — Method

Give centre of an orthogonal grid cell of the grid with cellsize, 
centre at grid origin and rotated at angle α counterclockwise.

PhaseRetrieval.constructSHPhaseDiversity — Method

constructSHPhaseDiversity(wfs::SHSensor, d, λ )

Construct SH phase diversity of minimal size for wfs within aperture of diameter d and wavelength λ. Return tuple of arrays for the aperture and phase.

PhaseRetrieval.diaphragm — Method

diaphragm(lens, diam)

Change the lens diameter of an ImagingLens or an ImagingSensor.

Diaphragm can be larger than the lens diameter (there is no "hardware limitation").

Example

julia> diaphragm(ImagingLens(300mm, 25mm), 10mm)
ImagingLens(0.3, 0.01)

PhaseRetrieval.diskmatrix — Function

Create array of zeroes with disk of ones of the relative diameter r

PhaseRetrieval.getstoragetype — Method

getstoragetype(cam::CameraChip)

Get the storage type returned by the camera cam.

PhaseRetrieval.intensity — Method

intensity(field)

Calculate intensity of the field.

Examples

julia> ap,_ = PhaseRetrieval.aperture(-1:.2:1,-1:.2:1,.8);

julia> imf = PhaseRetrieval.toimageplane(ap);

julia> PhaseRetrieval.intensity(imf) .|> round
11×11 Matrix{Float64}:
 18.0  3.0   1.0   2.0    0.0    1.0    0.0   2.0   1.0  3.0  18.0
  3.0  0.0   6.0   5.0    0.0    1.0    0.0   5.0   6.0  0.0   3.0
  1.0  6.0  11.0   4.0    1.0    5.0    1.0   4.0  11.0  6.0   1.0
  2.0  5.0   4.0   1.0   20.0   38.0   20.0   1.0   4.0  5.0   2.0
  0.0  0.0   1.0  20.0   78.0  118.0   78.0  20.0   1.0  0.0   0.0
  1.0  1.0   5.0  38.0  118.0  169.0  118.0  38.0   5.0  1.0   1.0
  0.0  0.0   1.0  20.0   78.0  118.0   78.0  20.0   1.0  0.0   0.0
  2.0  5.0   4.0   1.0   20.0   38.0   20.0   1.0   4.0  5.0   2.0
  1.0  6.0  11.0   4.0    1.0    5.0    1.0   4.0  11.0  6.0   1.0
  3.0  0.0   6.0   5.0    0.0    1.0    0.0   5.0   6.0  0.0   3.0
 18.0  3.0   1.0   2.0    0.0    1.0    0.0   2.0   1.0  3.0  18.0

PhaseRetrieval.lineararray — Function

lineararray(xrange, yrange, kx, ky, k0=0)

Evaluate values of linear function $kx⋅x + ky⋅y + k0$ on array $xrange × yrange$.

lineararray(size, kx, ky, k0=0)

Use 1:size as xrange and yrange.

lineararray(xrange, yrange, a::Vector, k0=0)

Use first and second components of vector a as kx and ky.

PhaseRetrieval.logrescale — Function

logrescale(array, α = 5)

Rescale in log-scale: maximum will correspond to 1, $10^{-α}$ will correspond to 0.

PhaseRetrieval.orthIndexes — Function

Generate array of indexes of an orthogonal geometry

PhaseRetrieval.phwrap — Method

Wrap Phase

PhaseRetrieval.psf — Method

psf(amplitude, phase) -> psfimage
psf(pupilfield) -> psfimage

Calculate psf for given amplitude and phase. Both PSF and pupil arrays are assumed to have the origin in their center pixel ((N+1)÷2 for dimension size N).

Examples

julia> ap,_ = PhaseRetrieval.aperture(-1:.2:1,-1:.2:1,.8); ap
11×11 Matrix{Float64}:
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  1.0  1.0  1.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  1.0  1.0  1.0  1.0  1.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  1.0  1.0  1.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0

julia> Int.(round.(psf(ap)))
11×11 Matrix{Int64}:
 18  3   1   2    0    1    0   2   1  3  18
  3  0   6   5    0    1    0   5   6  0   3
  1  6  11   4    1    5    1   4  11  6   1
  2  5   4   1   20   38   20   1   4  5   2
  0  0   1  20   78  118   78  20   1  0   0
  1  1   5  38  118  169  118  38   5  1   1
  0  0   1  20   78  118   78  20   1  0   0
  2  5   4   1   20   38   20   1   4  5   2
  1  6  11   4    1    5    1   4  11  6   1
  3  0   6   5    0    1    0   5   6  0   3
 18  3   1   2    0    1    0   2   1  3  18

julia> psf(ones(3,3),zeros(3,3))
3×3 Matrix{Float64}:
 0.0   0.0  0.0
 0.0  81.0  0.0
 0.0   0.0  0.0
 
julia> psf(ones(Complex{Float64},4,4))
4×4 Matrix{Float64}:
 0.0  0.0    0.0  0.0
 0.0  0.0    0.0  0.0
 0.0  0.0  256.0  0.0
 0.0  0.0    0.0  0.0

PhaseRetrieval.quadratic — Method

Generate quadratic array

PhaseRetrieval.rescale — Method

rescale(array)

Rescale array between 0 and 1

PhaseRetrieval.rescale_minmax — Method

rescale_minmax(array)

Rescale array between 0 and 1 and return the rescaled array and the min, max values

PhaseRetrieval.roi — Method

roi(cam, dims)

Create camera that represents centered regeion of interest (ROI) of cam.

ROI can be larger than the camera size! :-)

Arguments

cam: camera (CameraChip) or imaging sensor (ImagingSensor)

dims::Tuple{Int, Int}: new (width, height) of the image. Single number defines a square ROI.

PhaseRetrieval.slopearray — Function

slopearray(dom::CartesianDomain2D, dx, dy, k0=0) generates linear array defined on the domain such that
the total growth in x and y direction is dx and dy.

PhaseRetrieval.subdivide — Method

subdivide(arr, Q)

Subdivide array of dimension MQ x NQ in Q^2 stacked tiles of size M x N.

PhaseRetrieval.subdivide_sum — Method

subdivide_sum(arr,Q)

Divide array in quadratic cells of size Q × Q and sum the elements with the same indexes in each cell.

PhaseRetrieval.tile — Method

tile(arr, Q)

Divide array in quadratic cells of size Q × Q and stack them along the third dimension.

PhaseRetrieval.toimageplane — Method

toimageplane(field [, method = Fourier()])

Calculate the field in the focal plane from the field in the pupil plane using method.

Available methods:

Fourier()

ChirpZ – not implemented

MVM – not implemented

Index

PhaseRetrieval.camerasdict

PhaseRetrieval.lensesdict

PhaseRetrieval.storagetype

PhaseRetrieval.AbstractPRproblem

PhaseRetrieval.CameraChip

PhaseRetrieval.Fourier

PhaseRetrieval.GS

PhaseRetrieval.GSparam

PhaseRetrieval.ImagingLens

PhaseRetrieval.ImagingSensor

PhaseRetrieval.Linear

PhaseRetrieval.MLA

PhaseRetrieval.PDplan

PhaseRetrieval.PRproblem

PhaseRetrieval.Quadratic

PhaseRetrieval.SHSensor

PhaseRetrieval.SimConfig

PhaseRetrieval.hwConfig

AlternatingProjections.AbstractProblems.solve

PhaseRetrieval.SHdiversity

PhaseRetrieval.SHdiversity

PhaseRetrieval.appsftoPR

PhaseRetrieval.binning

PhaseRetrieval.centre

PhaseRetrieval.constructSHPhaseDiversity

PhaseRetrieval.diaphragm

PhaseRetrieval.diskmatrix

PhaseRetrieval.getstoragetype

PhaseRetrieval.intensity

PhaseRetrieval.lineararray

PhaseRetrieval.logrescale

PhaseRetrieval.orthIndexes

PhaseRetrieval.phwrap

PhaseRetrieval.psf

PhaseRetrieval.quadratic

PhaseRetrieval.rescale

PhaseRetrieval.rescale_minmax

PhaseRetrieval.roi

PhaseRetrieval.slopearray

PhaseRetrieval.subdivide

PhaseRetrieval.subdivide_sum

PhaseRetrieval.tile

PhaseRetrieval.toimageplane